Struct wikilib::Wiki

pub struct Wiki { /* fields omitted */ }

Global processing structure

Methods

impl Wiki

fn new() -> Self

Create a new Wiki instance

fn init_logging(&mut self, level: LogLevel) -> Result<()>

Creates a new instance of the processing lib

fn read_from_directory(&mut self, directory: &str) -> Result<()>

Reads all markdown files recursively from a given directory. Clears the current available paths

fn list_current_paths(&self)

Print absolute path of all added md files

fn read_content_from_current_paths(&self,
input_root_dir: &str,
output_directory: &str)
-> Result<()>

Read the content of all files and convert it to HTML

fn serve(&self, output_directory: &str) -> Result<()>

Create an HTTP server serving the generated files